Bug 583125 - Use IEC 61499 coordinate system in model: Positions
This first commit uses the IEC 61499 coordinate system for storing all
positions. For drawing the coordinates are converted to screen
coordinates on edit the coordinates are converted back.

+++ plugins/org.eclipse.fordiac.ide.fb.interpreter/src/org/eclipse/fordiac/ide/fb/interpreter/testappgen/TestEccGenerator.java
@@ -45,10 +45,10 @@ public class TestEccGenerator {
 
 		ecc.getECTransition().add(transition);
 
-		final int x1 = from.getPosition().getX();
-		final int x2 = to.getPosition().getX();
-		final int y1 = from.getPosition().getY();
-		final int y2 = to.getPosition().getY();
+		final double x1 = from.getPosition().getX();
+		final double x2 = to.getPosition().getX();
+		final double y1 = from.getPosition().getY();
+		final double y2 = to.getPosition().getY();
 		AbstractBlockGenerator.addPosition(transition, x1 + ((x2 - x1) / 2), y2 + ((y1 - y2) / 2));
 	}
